requirements.txt
Django==3.2.9
psycopg2-binary==2.9.5
djangorestframework==3.13.1
